app/requirements.txt
curl-cffi==0.5.5
beautifulsoup4==4.10.0
pyzmq==22.3.0